# Docs/openapi.yaml
openapi: 3.0.3
info:
  title: tracklist API
  description: |
    REST API для автентифікації, керування користувачами,
    медіа-контентом, відгуками, коментарями, списками та взаємодіями.
  version: 1.0.0
servers:
  - url: http://localhost:7288
    description: Local ASP.NET Core API
...
paths:
...
  /api/media/{mediaId}/translations:
    post:
      tags: [Media]
      summary: Додавання перекладу медіа
      parameters:
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaTranslationRequest'
            example:
              languageCode: uk
              title: Назва українською
              description: Опис українською
      responses:
        "201":
          description: Translation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaTranslation'
  /api/media/translations/{id}:
    put:
      tags: [Media]
      summary: Редагування перекладу медіа
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaTranslationRequest'
      responses:
        "200":
          description: Translation updated

    delete:
      tags: [Media]
      summary: Видалення перекладу медіа
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Translation deleted
...
  /api/Playlists/{playlistId}/media/{mediaId}:
    delete:
      tags: [Playlists]
      summary: Видалення медіа зі списку
      parameters:
        - name: playlistId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Media removed from playlist
...
components:
  schemas:
...
    MediaTranslation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mediaId:
          type: string
          format: uuid
        languageCode:
          type: string
          maxLength: 5
          example: uk
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        status:
          $ref: '#/components/schemas/TranslationStatus'
        processedByUserId:
          type: string
          format: uuid
          nullable: true
    MediaTranslationRequest:
      type: object
      properties:
        languageCode:
          type: string
        title:
          type: string
        description:
          type: string
...
    TranslationStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
...